Implement id3 frame parsing
Currently the parser works all the way up to any padding at the
end of the tag and then hard stops. This grabs all the frames
(I hope) but doesn't move up to the point where the mp3
file starts. If users ever want to write out their own tags,
figuring out where the mp3 player starts would be an internal
feature that would be necessary.